python_requirements.txt
django
dotenv
django-autocomplete-light
django-debug-toolbar
pyzotero
iso639
pylobid
django-htmx
django-haystack
whoosh
dominate
